Implement DateTrait across multiple entities
The DateTrait was created and used across multiple entities (Tricks, Picture, Comment) to handle date and time related properties and methods (such as createdAt, updatedAt, getCreatedAt, setCreatedAt, getUpdatedAt, setUpdatedAt and updateTimestamps). This refactor significantly reduces redundancy and improves the maintainability of the code.
